requirements.txt
pyspark
kafka-python
requests
nltk
streamlit
altair
